+++ lib/scroll.d.ts
@@ -1,0 +1,3 @@
+declare type ScrollCallback = () => any | void;
+declare function scroll(position: number | string, callback?: ScrollCallback): void;
+export default scroll;