+++ rollup.config.js
@@ -5,6 +5,7 @@ const def = {
   input: {
     index: 'src/index.ts',
     history: 'src/history.ts',
+    scroll: 'src/scroll.ts',
     decode: 'src/decode.ts',
     setSearch: 'src/setSearch.ts',
     removeSearch: 'src/removeSearch.ts',