+++ src/index.ts
@@ -2,6 +2,7 @@ import decode from './decode'
 import parseUrl from './parseUrl'
 import setSearch from './setSearch'
 import removeSearch from './removeSearch'
+import scroll from './scroll'
 
 export {default} from './history'
 export * from './history'
@@ -9,6 +10,7 @@ export * from './parseUrl'
 
 export {
   decode,
+  scroll,
   parseUrl,
   setSearch,
   removeSearch